providing examples for LLM
when we send some prompts to LLMs, they might get confused because a lot of reasons; for example the length of our prompt. to tackle this problem we should provide some examples about the things we want the model to do and know the way of our thinking. so in this module i have trained this object with providing 2 examples in Farsi and the model will answer more accurately.