Creating Memory
We have 3 ways of making a memory for our LLM. One way is to get all the prompts that model and user have generated and send all of them to our model as a new prompt to make a memory for it. The other way is that we give a summary of previous prompts to our model. The last way is that we choose a k number that we give the important things of those k prompts as a new prompt to our model. In this module, I have implemented all of these ways using LangChain